Adding project 2 parts 0,1 and 2
These hello.cpp files can be copied into the PX4 source folder and flashed to the FMU based on which part needs to be tested. For project 2 part 2, hello_p2.cpp needs to be flashed to the FMU and the python program CameraSensorCode.py needs to be run on the RPi.